move notify test out
