+++ archery/settings.py
@@ -86,6 +86,10 @@ AVAILABLE_ENGINES = {
     "phoenix": {"path": "sql.engines.phoenix:PhoenixEngine"},
     "odps": {"path": "sql.engines.odps:ODPSEngine"},
 }
+ENABLED_NOTIFIERS = (
+    "sql.notify:FeishuWebhookNotifier",
+    "sql.notify:GenericWebhookNotifier"
+)
 ENABLED_ENGINES = env("ENABLED_ENGINES")
 
 # Application definition
